@@ -109,22 +121,24 @@
     "postcss-flexbugs-fixes": "^4.1.0",
     "postcss-loader": "^3.0.0",
     "postcss-scss": "^2.0.0",
-    "postgraphile": "^4.3.2",
-    "prop-types": "^15.6.1",
-    "react": "^16.7.0",
-    "react-apollo": "^2.5.2",
-    "react-dom": "^16.7.0",
-    "react-helmet": "^5.2.0",
-    "react-hot-loader": "^4.6.5",
-    "react-redux": "^6.0.0",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "redux": "^4.0.1",
-    "redux-act": "^1.7.4",
+    "postgraphile": "^4.5.5",
+    "prop-types": "^15.7.2",
+    "react": "^16.12.0",
+    "react-apollo": "^3.1.3",
+    "react-dom": "^16.12.0",
+    "react-helmet": "^5.2.1",
+    "react-hot-loader": "^4.12.18",
+    "react-redux": "^7.1.3",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
+    "redux": "^4.0.5",
+    "redux-act": "^1.7.7",
     "redux-thunk": "^2.3.0",
     "rimraf": "^2.6.2",
     "sass-loader": "^7.1.0",
+    "slash": "^3.0.0",
     "style-loader": "^0.23.1",
+    "subscriptions-transport-ws": "^0.9.16",
     "uglifyjs-webpack-plugin": "^1.2.4",
     "upath": "^1.1.0",
     "webpack": "^4.29.0",
